refactor(apps/web): divy up toast lib
