+++ apps/web/package.json
@@ -63,7 +63,7 @@
 		"@vitejs/plugin-react": "^4.2.1",
 		"@vitest/coverage-v8": "1.4.0",
 		"autoprefixer": "^10.4.13",
-		"cypress": "^13.7.1",
+		"cypress": "^13.8.1",
 		"dotenv": "^16.4.5",
 		"husky": "^8.0.1",
 		"jsdom": "^22.1.0",
