+++ apps/web/src/lib/toast/types.ts
@@ -1,0 +1,4 @@
+export type ToastMessage = {
+	text: string;
+	type: string;
+};
